requirements.txt

streamlit
requests
pandas
feedparser
newsapi-python
python-dateutil
